--- src/parsers/mapped.ts
@@ -1,6 +1,0 @@
-import { Parser } from "./parser";
-
-export interface MappedParser {
-  name: string;
-  parser: Parser;
-}
