--- src/parsers/parser.ts
@@ -1,3 +1,0 @@
-export abstract class Parser {
-  abstract parse(payload: Record<string, unknown> | string): unknown;
-}
