--- src/parsers/pass_through.ts
@@ -1,7 +1,0 @@
-import { Parser } from "./parser";
-
-export class PassThroughParser extends Parser {
-  parse(payload: unknown): unknown {
-    return payload;
-  }
-}
